[Chore] Move `BaseRenderer` to `base.py`
