+++ vllm/renderers/__init__.py
@@ -1,8 +1,8 @@
 # SPDX-License-Identifier: Apache-2.0
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 
+from .base import BaseRenderer
 from .params import ChatParams, TokenizeParams, merge_kwargs
-from .protocol import BaseRenderer
 from .registry import RendererRegistry, renderer_from_config
 
 __all__ = [
