+++ vllm/renderers/base.py
